Securite: extrait la règle "nombre_dangers_bien_identifies_avant_aide_1"
