+++ app/models/restitution/metriques_helper.rb
@@ -2,7 +2,15 @@
 
 module Restitution
   class MetriquesHelper
+    EVENEMENT = {
+      ACTIVATION_AIDE_1: 'activationAide'
+    }.freeze
+
     class << self
+      def premier_evenement_du_nom(evenements, nom_evenement)
+        evenements.find { |e| e.nom == nom_evenement }
+      end
+
       def temps_entre_couples(evenements)
         les_temps = []
         evenements.each_slice(2) do |e1, e2|
